rename: 임베디드 타입 디렉토리 정리
