+++ src/main/java/com/example/gojipserver/domain/checklist/entity/bathroomstatus/ShowerHead.java
@@ -1,0 +1,6 @@
+package com.example.gojipserver.domain.checklist.entity.bathroomstatus;
+
+public enum ShowerHead {
+
+    GOOD, BAD
+}
