+++ src/main/java/com/example/gojipserver/domain/checklist/entity/bathroomstatus/Tile.java
@@ -1,0 +1,6 @@
+package com.example.gojipserver.domain.checklist.entity.bathroomstatus;
+
+public enum Tile {
+
+    GOOD, BAD
+}
