+++ src/main/java/com/example/gojipserver/domain/checklist/entity/bathroomstatus/Toilet.java
@@ -1,0 +1,6 @@
+package com.example.gojipserver.domain.checklist.entity.bathroomstatus;
+
+public enum Toilet {
+
+    GOOD, BAD
+}
