+++ src/main/java/com/example/gojipserver/domain/checklist/entity/cost/Cost.java
@@ -1,4 +1,4 @@
-package com.example.gojipserver.domain.checklist.entity;
+package com.example.gojipserver.domain.checklist.entity.cost;
 
 import jakarta.persistence.Embeddable;
 import jakarta.persistence.Embedded;
