+++ src/main/java/com/example/gojipserver/domain/checklist/entity/option/InnerOption.java
@@ -1,4 +1,4 @@
-package com.example.gojipserver.domain.checklist.entity;
+package com.example.gojipserver.domain.checklist.entity.option;
 
 import jakarta.persistence.Embeddable;
 import lombok.AccessLevel;
