+++ src/main/java/com/example/gojipserver/domain/checklist/entity/roomcondition/Building.java
@@ -1,0 +1,6 @@
+package com.example.gojipserver.domain.checklist.entity.roomcondition;
+
+public enum Building {
+
+    GOOD, BAD
+}
