+++ src/main/java/com/example/gojipserver/domain/checklist/entity/roomstatus/Boiler.java
@@ -1,0 +1,7 @@
+package com.example.gojipserver.domain.checklist.entity.roomstatus;
+
+public enum Boiler {
+
+
+    INDIVIDUAL, COMMON //개별, 공동
+}
