--- src/main/java/com/example/gojipserver/domain/checklist/entity/status/Boiler.java
@@ -1,7 +1,0 @@
-package com.example.gojipserver.domain.checklist.entity.status;
-
-public enum Boiler {
-
-
-    INDIVIDUAL, COMMON //개별, 공동
-}
