--- src/main/java/com/example/gojipserver/domain/checklist/entity/status/Sink.java
@@ -1,6 +1,0 @@
-package com.example.gojipserver.domain.checklist.entity.status;
-
-public enum Sink {
-
-    GOOD, BAD
-}
