--- src/main/java/com/example/gojipserver/domain/checklist/entity/status/Toilet.java
@@ -1,6 +1,0 @@
-package com.example.gojipserver.domain.checklist.entity.status;
-
-public enum Toilet {
-
-    GOOD, BAD
-}
